+++ tests/client/stores/fromError.js
@@ -1,4 +1,0 @@
-module.exports = {
-  name : 'fromError',
-  from : '@fakeStore'
-};