+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.xx.x
 - Features:
   - Add `errors` attribute in `lunaris.http` response when an error is detected. The store `lunarisErrors` has the server errors in `messageErrorServer` attribute which is an object as `{ error : Int | String, message : Sring, errors : Array }`.
-  - Add `from` attribute in store object to inherit from a store its map and collection.
+  - Add `inherits` attribute in store object to inherit from a store its map.
 - Fix:
   - Fix store setPagination method using wrong offset